Refactor Classification components
* Extract CustomClassificationWidget.jsx, CustomGroupsWidget.jsx and CustomProbabilityWidget.jsx from ClassificationForm.jsx
* Extract `addNodePaths`, `allowedClasses` and `useClassifications` functions into helpers.js
