+++ static/js/components/candidate/AddClassificationsScanningPage.jsx
@@ -11,8 +11,8 @@ import IconButton from "@mui/material/IconButton";
 import Tooltip from "@mui/material/Tooltip";
 import ClassificationSelect from "../classification/ClassificationSelect";
 import * as Actions from "../../ducks/source";
-import { allowedClasses } from "../classification/ClassificationForm";
 import Button from "../Button";
+import { allowedClasses } from "../../utils/helpers";
 
 const AddClassificationsScanningPage = ({ obj_id }) => {
   const [dialogOpen, setDialogOpen] = useState(false);
